move renaming of iterable to iterableLike in logic to 0.19.0
and remove String.Companion.format in testUtils
