+++ misc/specs/atrium-specs-common/src/main/kotlin/ch/tutteli/atrium/specs/integration/CharSequenceToContainSpecBase.kt
@@ -1,6 +1,6 @@
 package ch.tutteli.atrium.specs.integration
 
-import ch.tutteli.atrium.specs.format
+import ch.tutteli.atrium.core.polyfills.format
 import ch.tutteli.atrium.specs.lineSeparator
 import ch.tutteli.atrium.translations.DescriptionCharSequenceExpectation
 import org.spekframework.spek2.Spek
@@ -13,8 +13,7 @@ abstract class CharSequenceToContainSpecBase(spec: Root.() -> Unit) : Spek(spec)
         const val helloWorld = "Hello World, I am Oskar"
 
         val toContainDescr = DescriptionCharSequenceExpectation.TO_CONTAIN.getDefault()
-        val toContainIgnoringCase = String.format(
-            DescriptionCharSequenceExpectation.IGNORING_CASE.getDefault(),
+        val toContainIgnoringCase = DescriptionCharSequenceExpectation.IGNORING_CASE.getDefault().format(
             DescriptionCharSequenceExpectation.TO_CONTAIN.getDefault()
         )
         val numberOfOccurrences = DescriptionCharSequenceExpectation.NUMBER_OF_MATCHES.getDefault()
